feat: move remaining system and home configuration to modules
Move the remaining system and home configuration into respective
modules.

The home-manager my-desktop module parametises the previous
configuration allowing for simple configuration especially when it comes
to themes.

The nixos my-system module parametises the previous configuration aiming
for sensible parametisation to configure what is needed for a full
system bootstrap compared to a configuration based on already existing
system.